chore: refactor logging
