+++ pom.xml
@@ -23,6 +23,7 @@
                 <configuration>
                     <excludes>
                         <exclude>**/constant/**</exclude>
+                        <exclude>**/utils/**</exclude>
 
                         <exclude>**/client/openapi/OpenApiStub*</exclude>
                         <exclude>**/client/Client*</exclude>
