+++ src/main/kotlin/com/expediagroup/sdk/core/client/openapi/OpenApiStub.kt
@@ -17,10 +17,10 @@ package com.expediagroup.sdk.core.client.openapi
 
 import com.expediagroup.sdk.core.client.Client
 import com.expediagroup.sdk.core.configuration.ClientConfiguration
-import com.expediagroup.sdk.core.constant.Logging.RESPONSE_PAYLOAD_RECEPTION_FAILURE
-import com.expediagroup.sdk.core.constant.Logging.responseUnsuccessful
+import com.expediagroup.sdk.core.constant.LoggingMessage.RESPONSE_PAYLOAD_RECEPTION_FAILURE
 import com.expediagroup.sdk.core.model.error.Error
 import com.expediagroup.sdk.core.model.exception.ServiceException
+import com.expediagroup.sdk.core.utils.LoggingMessageProvider.responseUnsuccessful
 import io.ktor.client.call.body
 import io.ktor.client.engine.okhttp.OkHttp
 import io.ktor.client.statement.HttpResponse
