+++ src/main/kotlin/com/expediagroup/sdk/core/plugin/authentication/AuthenticationPlugin.kt
@@ -18,15 +18,15 @@ package com.expediagroup.sdk.core.plugin.authentication
 import com.expediagroup.sdk.core.configuration.Credentials
 import com.expediagroup.sdk.core.constant.ExceptionMessage.AUTHENTICATION_FAILURE
 import com.expediagroup.sdk.core.constant.Header
-import com.expediagroup.sdk.core.constant.Logging.TOKEN_CLEARING_IN_PROCESS
-import com.expediagroup.sdk.core.constant.Logging.TOKEN_CLEARING_SUCCESSFUL
-import com.expediagroup.sdk.core.constant.Logging.TOKEN_RENEWAL_FAILED
-import com.expediagroup.sdk.core.constant.Logging.TOKEN_RENEWAL_IN_PROCESS
-import com.expediagroup.sdk.core.constant.Logging.TOKEN_RENEWAL_SUCCESSFUL
-import com.expediagroup.sdk.core.constant.Logging.responseStatusCode
-import com.expediagroup.sdk.core.constant.Logging.tokenExpiresIn
+import com.expediagroup.sdk.core.constant.LoggingMessage.TOKEN_CLEARING_IN_PROCESS
+import com.expediagroup.sdk.core.constant.LoggingMessage.TOKEN_CLEARING_SUCCESSFUL
+import com.expediagroup.sdk.core.constant.LoggingMessage.TOKEN_RENEWAL_FAILED
+import com.expediagroup.sdk.core.constant.LoggingMessage.TOKEN_RENEWAL_IN_PROCESS
+import com.expediagroup.sdk.core.constant.LoggingMessage.TOKEN_RENEWAL_SUCCESSFUL
 import com.expediagroup.sdk.core.model.exception.ClientException
 import com.expediagroup.sdk.core.plugin.Plugin
+import com.expediagroup.sdk.core.utils.LoggingMessageProvider.responseStatusCode
+import com.expediagroup.sdk.core.utils.LoggingMessageProvider.tokenExpiresIn
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.plugins.auth.Auth
